# frontend/scripts/openapi/contracts/auth.json
{
  "components": {
    "schemas": {
...
      "DatabasePlaybookCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "anyOf": [
              {
                "maxLength": 1048576,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "description": {
            "default": "",
            "maxLength": 1024,
            "title": "Description",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "entrypoint": {
            "anyOf": [
              {
                "maxLength": 512,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entrypoint"
          },
          "files": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PlaybookFileRequest"
                },
                "maxItems": 256,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Files"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "parameterSchema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parameterschema"
          },
          "supportsCheckMode": {
            "default": false,
            "title": "Supportscheckmode",
            "type": "boolean"
          }
        },
        "required": [
          "name"
        ],
        "title": "DatabasePlaybookCreateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlaybookRestoreRequest": {
        "additionalProperties": false,
        "properties": {
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "title": "PlaybookRestoreRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpsComposer API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/playbooks/database": {
      "post": {
        "operationId": "createDatabasePlaybook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabasePlaybookCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabasePlaybookDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Database Playbook",
        "tags": [
          "playbooks"
        ]
      }
    },
    "/api/v1/playbooks/database/import": {
      "post": {
        "operationId": "importPlaybookProjectZip",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_importPlaybookProjectZip"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybookZipImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Playbook Project Zip",
        "tags": [
          "playbooks"
        ]
      }
    },
    "/api/v1/playbooks/database/{playbook_id}": {
      "delete": {
        "operationId": "deleteDatabasePlaybook",
        "parameters": [
          {
            "in": "path",
            "name": "playbook_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Playbook Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Database Playbook",
        "tags": [
          "playbooks"
        ]
      },
      "get": {
        "operationId": "getDatabasePlaybook",
        "parameters": [
          {
            "in": "path",
            "name": "playbook_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Playbook Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabasePlaybookDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Database Playbook",
        "tags": [
          "playbooks"
        ]
      },
      "put": {
        "operationId": "updateDatabasePlaybook",
        "parameters": [
          {
            "in": "path",
            "name": "playbook_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Playbook Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatabasePlaybookUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabasePlaybookDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Database Playbook",
        "tags": [
          "playbooks"
        ]
      }
    },
    "/api/v1/playbooks/database/{playbook_id}/revisions": {
      "get": {
        "operationId": "listPlaybookRevisions",
        "parameters": [
          {
            "in": "path",
            "name": "playbook_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Playbook Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlaybookRevisionResponse"
                  },
                  "title": "Response Listplaybookrevisions",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Playbook Revisions",
        "tags": [
          "playbooks"
        ]
      }
    },
    "/api/v1/playbooks/database/{playbook_id}/revisions/{revision}": {
      "get": {
        "operationId": "getPlaybookRevision",
        "parameters": [
          {
            "in": "path",
            "name": "playbook_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Playbook Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "title": "Revision",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybookRevisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Playbook Revision",
        "tags": [
          "playbooks"
        ]
      }
    },
    "/api/v1/playbooks/database/{playbook_id}/revisions/{revision}/diff": {
      "get": {
        "operationId": "diffPlaybookRevisions",
        "parameters": [
          {
            "in": "path",
            "name": "playbook_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Playbook Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "title": "Revision",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "againstRevision",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Againstrevision",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybookDiffResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Diff Playbook Revisions",
        "tags": [
          "playbooks"
        ]
      }
    },
    "/api/v1/playbooks/database/{playbook_id}/revisions/{revision}/export": {
      "get": {
        "operationId": "exportPlaybookProjectZip",
        "parameters": [
          {
            "in": "path",
            "name": "playbook_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Playbook Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "title": "Revision",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Playbook Project Zip",
        "tags": [
          "playbooks"
        ]
      }
    },
    "/api/v1/playbooks/database/{playbook_id}/revisions/{revision}/restore": {
      "post": {
        "operationId": "restorePlaybookRevision",
        "parameters": [
          {
            "in": "path",
            "name": "playbook_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Playbook Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "title": "Revision",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaybookRestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabasePlaybookDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Playbook Revision",
        "tags": [
          "playbooks"
        ]
      }
    },
    "/api/v1/playbooks/detail": {
      "get": {
        "operationId": "getMountedPlaybook",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybookSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Mounted Playbook",
        "tags": [
          "playbooks"
        ]
      }
    },
    "/api/v1/playbooks/validate": {
      "post": {
        "operationId": "validatePlaybook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaybookValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybookValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Playbook",
        "tags": [
          "playbooks"
        ]
      }
    },
...
  }
}